refactor: move generic classes to separate folder
