+++ src/M31.FluentApi.Generator/SourceGenerators/FluentApiClassInfo.cs
@@ -1,4 +1,5 @@
 using M31.FluentApi.Generator.Commons;
+using M31.FluentApi.Generator.SourceGenerators.Generics;
 
 namespace M31.FluentApi.Generator.SourceGenerators;
