+++ src/M31.FluentApi.Generator/SourceGenerators/Generics/GenericInfo.cs
@@ -1,6 +1,6 @@
 using M31.FluentApi.Generator.Commons;
 
-namespace M31.FluentApi.Generator.SourceGenerators;
+namespace M31.FluentApi.Generator.SourceGenerators.Generics;
 
 internal class GenericInfo
 {
